src/requirements/common.txt
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.4
requests==2.22.0
celery==4.4.0
redis==3.4.1
django-import-export==2.0.2
